requirements.txt
torch==1.0.0
scipy
argparse
Pillow
sacred
easydict
pymongo
pyyaml
matplotlib
pyrender
triangle
